Remove `@Sharable` and replace it with default method
Motivation:

Now that we depend on java8+ we should just use a default method that can be override to declare a handler sharable. This makes things a bit cheaper and also allow to fail to restrict sharable when compiling by let implementations declare the method final.

Modifications:

- Remove annotation and add default method to ChannelHandler interface
- Adjust implementations

Result:

Better way to declare a handler sharable or not